plugin: new vim-function ZemGetMatches(query, limit=None)
returns an Array of dict with all matches of the query.

Allows for example to include the header of a function/define, by
querying for define or prototyp, requesting and taking only the first
match, using the file attribute and removing the path to the file, and
then storing it in a register to insert an include directive.

nnoremap <leader>zh :let @h=substitute(ZemGetMatches("=Proto =Def <C-R><C-W>",1)[0]['file'],'.*/','','')<CR>?^#include<CR>:nohlsearch<CR>o<CR>#include "<C-R>h"<ESC>0